+++ actix_web_2023/src/main.rs
@@ -1,4 +1,4 @@
-use actix_web_2023::app_run;
+use actix_web_2023::router::app_run;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
